agents/requirements.txt
# ============================================================================
# Firebase Functions 2nd Gen SDK
# ============================================================================
firebase-functions>=0.4.0
firebase-admin>=6.5.0

# ============================================================================
# Core Dependencies
# ============================================================================
flask==3.1.0

# ============================================================================
# Anthropic Claude API
# ============================================================================
anthropic>=0.72.0

# ============================================================================
# Configuration & Validation
# ============================================================================
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# ============================================================================
# External API Calls
# ============================================================================
requests>=2.31.0